requirements.txt
pandas
numpy
scikit-learn
statsmodels
matplotlib
seaborn
plotly
streamlit
sqlalchemy
psycopg2-binary
pymysql
fastapi
uvicorn
python-pptx
